Add config menu bar, motor ang rate widget, remove old tkinter gui and create cli.
